+++ R/tax_gbif_occur_pq.R
@@ -23,6 +23,11 @@
 #' @param time_to_sleep (numeric, default 0.3) Time to sleep between two calls to
 #'  rgbif::occ_search(). Useful to avoid to be blocked by GBIF. Try to increase
 #'  this value if you are blocked by the error "To download GBIF occurrence data in bulk, please request..."
+#' @param discard_genus_alone (logical, default `TRUE` when
+#'  `taxonomic_rank == "currentCanonicalSimple"`). Passed to
+#'  [taxonomic_rank_to_taxnames()].
+#' @param discard_NA (logical, default `TRUE`). Passed to
+#'  [taxonomic_rank_to_taxnames()].
 #'
 #' @returns Either a tibble (if add_to_phyloseq = FALSE) or a new phyloseq
 #'  object, if add_to_phyloseq = TRUE, with new column(s) in the tax_table.
@@ -64,7 +69,7 @@ tax_gbif_occur_pq <- function(
   by_years = FALSE,
   verbose = TRUE,
   time_to_sleep = 0.3,
-  discard_genus_alone = taxonomic_rank == "currentCanonicalSimple",
+  discard_genus_alone = identical(taxonomic_rank, "currentCanonicalSimple"),
   discard_NA = TRUE
 ) {
   if (!is.null(taxnames) && !is.null(physeq)) {